% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{analysis_functions}
\alias{analysis_functions}
\alias{AggregateTemplate}
\title{Analysis functions for EROSBIT dataset}
\usage{
AggregateTemplate(BITraster, FUN)
}
\description{
NA
}
\details{
NA
}
\author{
Brandon McNellis
}
