% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBIT.R
\name{GetBIT}
\alias{GetBIT}
\title{Collects BIT data specified by a BITraster-class object}
\usage{
GetBIT(BITraster)
}
\arguments{
\item{BITraster}{Object of class BITraster}
}
\description{
NA
}
\author{
Brandon McNellis
}
