% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper_functions}
\alias{helper_functions}
\alias{VarNames}
\alias{AllYears}
\alias{matchResolution}
\title{Miscellaneous supporting functions specific to this package}
\usage{
VarNames()

AllYears()

matchResolution(x, ref, method = "bilinear", filename = "", ...)
}
\description{
NA
}
\author{
Brandon McNellis
}
