% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_BIT}
\alias{sample_BIT}
\title{Sample raster of BIT cover}
\format{
A RasterLayer with 333 rows, 333 columns, and 110,889 pixels
}
\usage{
sample_BIT
}
\description{
Raster is generated from 1985 Annual Herbaceous Cover in the central
Colorado Plateau. Resolution is 30m x 30m.
}
\keyword{datasets}
