% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_mask}
\alias{sample_mask}
\title{Sample mask of BIT cover}
\format{
A RasterLayer with 333 rows, 333 columns, and 110,889 pixels
}
\usage{
sample_mask
}
\description{
Mask is generated from `sample_BIT`.
}
\details{
1 = >15% annual herbaceous cover, 0 = <= 15% annual herbaceous cover.
}
\keyword{datasets}
