% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BITraster.R
\docType{class}
\name{BITraster}
\alias{BITraster}
\alias{BITraster-class}
\alias{initialize,BITraster-method}
\alias{ImportBIT}
\alias{.subset_BIT_points}
\alias{SubsetPoints,ANY,matrix-method}
\alias{SubsetPoints,ANY,data.frame-method}
\alias{SubsetPoints,ANY,SpatialPoints-method}
\title{S4 class for extending Raster* to EROS BIT dataset}
\usage{
\S4method{initialize}{BITraster}(.Object, ...)

ImportBIT(dir, variables, years)

.subset_BIT_points(BITraster, points, variables, years, parallel = FALSE)

\S4method{SubsetPoints}{ANY,matrix}(BITraster, points, variables, years, parallel = FALSE)

\S4method{SubsetPoints}{ANY,data.frame}(BITraster, points, variables, years, parallel = FALSE)

\S4method{SubsetPoints}{ANY,SpatialPoints}(BITraster, points, variables, years, parallel = FALSE)
}
\description{
NA
}
\details{
NA
}
\section{Slots}{

\describe{
\item{\code{variables}}{Character vector describing the contained BIT variables}

\item{\code{data_directory}}{Full path to un-zipped BIT data, defaults to current working directory}

\item{\code{temp_directory}}{Optional place to put temporary raster files (they get big)}

\item{\code{years}}{Years of the analysis}
}}

\author{
Brandon McNellis
}
